+++ pulsar/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -
 '''Event driven concurrent framework for Python'''
-VERSION = (0, 6, 1, 'beta', 1)
+VERSION = (0, 6, 1, 'alpha', 2)
 
 from .utils.version import get_version