+++ pulsar/apps/test/case.py
@@ -29,12 +29,8 @@ else: # pragma nocover
         mock = None
 
 
-__all__ = ['sequential', 'unittest', 'mock', 'get_stream']
+__all__ = ['sequential', 'unittest', 'mock']
 
-
-def get_stream(cfg):
-    r = unittest.TextTestRunner()
-    return r.stream
 
 def sequential(cls):
     '''Decorator for a :class:`TestCase` which cause its test functions to run