+++ pulsar/utils/log.py
@@ -19,17 +19,6 @@ from .structures import AttributeDictionary
 
 NOLOG = 100
 
-__all__ = ['dictConfig',
-           'NullHandler',
-           'process_global',
-           'LogginMixin',
-           'Synchronized',
-           'local_method',
-           'local_property',
-           'threadsafe',
-           'LocalMixin']
-
-
 LOG_LEVELS = {
     "critical": logging.CRITICAL,
     "error": logging.ERROR,
@@ -95,7 +84,6 @@ def update_config(config, c):
         if name in c:
             config[name].update(c[name])
 
-    
 def local_method(f):
     name = f.__name__
     def _(self):
@@ -116,6 +104,23 @@ def threadsafe(f):
     _.__doc__ = f.__doc__
     _.__name__ = f.__name__
     return _
+    
+    
+class WritelnDecorator(object):
+    """Used to decorate file-like objects with a handy 'writeln' method.
+    taken from python."""
+    def __init__(self,stream):
+        self.stream = stream
+
+    def __getattr__(self, attr):
+        if attr in ('stream', '__getstate__'):
+            raise AttributeError(attr)
+        return getattr(self.stream, attr)
+
+    def writeln(self, arg=None):
+        if arg:
+            self.write(arg)
+        self.write('\n') # text-mode streams translate to \r\n if needed
     
     
 class LocalMixin(object):