+++ pulsar/utils/tools/colors.py
@@ -6,15 +6,16 @@ with ANSI color codes for the appropriate color.
 
 For example, to print some text as green on supporting terminals::
 
-    from pulsar.utils.colors import green
+    from pulsar.utils.tools.colors import green
 
     print(green("This text is green!"))
 
 Because these functions simply return modified strings, you can nest them::
 
-    from fabric.colors import red, green
+    from pulsar.utils.tools.colors import red, green
 
-    print(red("This sentence is red, except for " + green("these words, which are green") + "."))
+    print(red("This sentence is red, except for " +
+        green("these words, which are green") + "."))
 
 If ``bold`` is set to ``True``, the ANSI flag for bolding will be flipped on
 for that particular invocation, which usually shows up as a bold or brighter
@@ -38,13 +39,13 @@ def _wrap_with(name,code):
     return inner
 
 _COLOR_MAP = {}
-red = _wrap_with('red','31')
-green = _wrap_with('green','32')
-yellow = _wrap_with('yellow','33')
-blue = _wrap_with('blue','34')
-magenta = _wrap_with('magenta','35')
-cyan = _wrap_with('cyan','36')
-white = _wrap_with('white','37')
+red = _wrap_with('red', '31')
+green = _wrap_with('green', '32')
+yellow = _wrap_with('yellow', '33')
+blue = _wrap_with('blue', '34')
+magenta = _wrap_with('magenta', '35')
+cyan = _wrap_with('cyan', '36')
+white = _wrap_with('white', '37')
 
 if platform.is_windows:  #pragma    nocover
     ColorFormatter = logging.Formatter